Changed database model, view profile info, view,signup, urls,
Changed the model and database to accept accountID so that the user now logs in with account id instead of email, it is unique. Added events and societies to show on account info except events is not yet fully completed need event relation in the model. Changed so that the account is verified as an owner from account id. Signup now has accountID as well. Finally view is logging in with account ID.