Gitignore on pyc files, removed serializer file, Modified models, society, urls and views
Removed the file serializer moved any functions into each file. Modified the models by adding members to keep track and count the number of users in societies when joining and leaving. Created functions for joining society, leaving, updating the society, and deleting society. Also fixed and sorted out the serializers for these functions and the get society details. Finally, added few paths for the functions in society and fixed the paths into using society name instead of id. Modified the views or removing the import serializers from the serializer file.