Modified society and permissions files
In the permissions, added IsAdmin (Which is same as the first but just left it, just to avoid any disruptions on merge), but the difference with that it has an error for if user isn't admin or accessing paths that admins are only allowed to access. Also, modified the society to handle specific errors like UpdateError, SocietyError, DeleteError etc (Don't know if it makes any difference but, personally it just makes it easier to find out what error).