Updated models to include course choices, year choices, interest tags, address and date of birth for account, friends model for friend relation.
Added a model for friends and a function to check if 2 accounts are friends in order to show events and societies for the friends profile, added tags and date of birth and address for account plus course choices and year choices. In profile updated everything to include the changed models and so that certain fields are shown to certain users.