Added upload profile pic and a list of members in the society
Managed to allow users to upload picture and can see that on their stats (tested using postman). Along with that, added a list of members in society, where if a user joins a society, it is added to the list of members (also shows the adminstatus to see if they're society admin or not).